chore: rename and refactor and clean
